test: refactored model tests, splitting into two categories
